Move the SOC-dependent functions to another class
