+++ moirae/models/components/__init__.py
@@ -1,0 +1,1 @@
+"""Definitions of components used by multiple types of models"""
