+++ moirae/models/ecm/utils.py
@@ -1,165 +1,6 @@
-"""Functions used in multiple components"""
-from typing import List, Optional, Union, Literal, Callable, Iterator
-from numbers import Number
+from typing import Union
 
 import numpy as np
-from pydantic import Field, PrivateAttr
-from scipy.interpolate import interp1d
-from numpy.polynomial.legendre import Legendre
-
-from moirae.models.base import HealthVariable, ListParameter, NumpyType
-
-
-class SOCInterpolatedHealth(HealthVariable):
-    """Health variables which vary as a function of SOC"""
-    base_values: ListParameter = \
-        Field(default=0, description='Values at specified SOCs')
-    soc_pinpoints: Optional[NumpyType] = Field(default=None, description='SOC pinpoints for interpolation.')
-    interpolation_style: Literal['linear', 'nearest', 'nearest-up', 'zero', 'slinear',
-                                 'quadratic', 'cubic', 'previous', 'next'] = \
-        Field(default='linear', description='Type of interpolation to perform')
-
-    # Internal caches
-    _ppoly_cache: tuple[np.ndarray, np.ndarray, Callable] | None = PrivateAttr(None)
-    """Cache for the interpolation function, and the interpolation points/soc points at which it was produced"""
-
-    def iter_parameters(self, updatable_only: bool = True, recurse: bool = True) -> Iterator[tuple[str, np.ndarray]]:
-        for name, param in super().iter_parameters(updatable_only, recurse):
-            if name != "soc_pinpoints":
-                yield name, param
-
-    def _get_function(self) -> Callable:
-        """Retrieve the interpolation function
-
-        Uses the internal cache, :attr:`_ppoly_cache`, if the :attr:`base_values` has not changed.
-
-        Returns:
-            The polynomial of interest
-        """
-
-        # Return the cached spline if it exists and the base values have not changed
-        if self._ppoly_cache is not None:
-            org_soc, orig_array, ppoly = self._ppoly_cache
-            if np.array_equal(orig_array, self.base_values) \
-                    and np.array_equal(org_soc, self.soc_pinpoints):
-                return ppoly
-
-            # Otherwise, clear the cache and continue
-            self._ppoly_cache = None
-
-        # Generate the SOC interpolation points if they don't already exist
-        if self.soc_pinpoints is None:
-            self.soc_pinpoints = np.linspace(0, 1, self.base_values.shape[-1])
-
-        # Make the spline then cache it
-        func = interp1d(self.soc_pinpoints,
-                        self.base_values,
-                        kind=self.interpolation_style,
-                        bounds_error=False,
-                        fill_value='extrapolate')
-        if self.interpolation_style not in {'linear', 'nearest', 'nearest-up'}:  # Don't cache lazy models
-            self._ppoly_cache = (self.soc_pinpoints.copy(), self.base_values.copy(), func)
-        return func
-
-    # TODO (wardlt): We might only use the diagonal terms from the output, maybe simplify by not computing off-diagonal
-    def get_value(self, soc: Union[Number, List, np.ndarray]) -> np.ndarray:
-        """
-        Computes value(s) at given SOC(s).
-
-        This function always returns a 3D array, of shape `(internal_batch_size, soc_batch_size, soc_dim)`, where
-        `internal_batch_size` is the batch size of the underlying health variable, `soc_batch_size` is the batch size
-        of the SOC array, and `soc_dim` is the dimensionality of the SOC. The SOC must be passed as either:
-        1. a 2D array of shape `(soc_batch_size, soc_dim)`
-        2. a 1D array of shape `(soc_dim,)`, in which case we will consider the `soc_batch_size` to be equal to 1
-        3. a 0D array (that is, a number), in which case both `soc_batch_size` and `soc_dim` are equal to 1.
-
-        Args:
-            soc: Values at which to compute the property
-        Returns:
-            Interpolated values as a 3D with dimensions (batch_size, soc_batch_size, soc_points)
-        """
-        # Determine which case we're dealing with
-        soc = self._adjust_soc_shape(soc)
-        internal_batch_size = self.base_values.shape[0]
-        soc_batch_size, soc_dim = soc.shape
-
-        # Special case: no interpolation
-        if self.base_values.shape[-1] == 1:
-            y = self.base_values[:, 0].copy()[:, None]  # shape = (internal_batch_size, 1)
-            y = np.tile(y, (soc_batch_size, 1, soc_dim))  # shape = (soc_batch, internal_batch, soc_dim)
-            y = np.swapaxes(y, 0, 1)  # shape = (internal_batch, soc_batch, soc_dim)
-        # Otherwise, run the interpolator, but the results mean something different
-        else:
-            y = self._get_function()(soc)  # interpolator adds batch dimension:
-            # If the SOC was batched, the shape is (internal_batch, soc_batch, soc_dim)
-            # Otherwise, the shape is (internal_batch, soc_dim)
-            y = y.reshape((internal_batch_size, soc_batch_size, soc_dim))
-
-        # Now, the y array has shape (internal_batch, soc_batch, soc_dim).
-        return y
-
-    def _adjust_soc_shape(self, soc: Union[Number, List, np.ndarray]) -> np.ndarray:
-        """Adjust the SOC from a user-provided shape to 2D array
-
-        Args:
-            soc: SOC, as provided by user
-        Returns:
-            2D array where first dimension is the batch and second is the SOC values.
-        """
-        soc = np.asarray(soc)  # Ensure it's an array, but don't copy it
-        if soc.ndim == 0:
-            return soc[None, None]
-        elif soc.ndim == 1:
-            return soc[None, :]
-        elif soc.ndim == 2:
-            return soc
-        else:
-            raise ValueError(f'SOC must be passed as at most a 2D array, but has shape {soc.shape}!')
-
-
-class ScaledSOCInterpolatedHealth(SOCInterpolatedHealth):
-    """SOC interpolated health with scaling factors to adjust the shape of the curve
-
-    Scaling factors adjust the shape produced from the interpolation by an additive
-    or multiplicative factor which varies as a function of SOC.
-    The scaling factor is described using
-    `Legendre polynomials <https://en.m.wikipedia.org/wiki/Legendre_polynomials>`_
-    defined over the region of [0, 1].
-    """
-
-    scaling_coeffs: ListParameter
-    """Coefficients of a Legendre polynomial used to adjust the interpolation"""
-    additive: bool = True
-    """Whether to add or multiply interpolated value with the scaling factor"""
-
-    def get_value(self, soc: Union[Number, List, np.ndarray]) -> np.ndarray:
-        # Evaluate the interpolated values
-        soc = self._adjust_soc_shape(soc)
-        interpolated = super().get_value(soc)
-
-        # Determine batch size of the output, prepare outputs
-        scale_batch = self.scaling_coeffs.shape[0]
-        inter_batch = self.base_values.shape[0]
-        soc_batch = soc.shape[0]
-
-        batch_size = max(scale_batch, inter_batch)
-        if inter_batch != batch_size:
-            assert inter_batch == 1, 'Inter batch should be either 1 or equal to the batch size'
-            output = np.tile(interpolated, (batch_size, 1, 1))
-        else:
-            output = interpolated
-
-        # Apply the scaling factors for each batch member
-        for b in range(batch_size):
-            # % is a shortcut which gets either the correct batch index for batched data,
-            #  and 1 for data which are not batched. It works
-            scaler = Legendre(coef=self.scaling_coeffs[b % scale_batch, :])
-            scaling_amount = scaler(soc[b % soc_batch, :])
-            if self.additive:
-                output[b, b % soc_batch, :] += scaling_amount
-            else:
-                output[b, b % soc_batch, :] *= 1 + scaling_amount
-        return output
 
 
 def realistic_fake_ocv(
